Add Rockstor 4 installer download place-holder links #7
As we are now at release candidate 8 (4.0.7-0) it makes sense to
broaden our surfacing of the Rockstor 4 DIY installer option. Here we
also introduce the new Rockstor 4 ARM64 compatibility as this may inform
folks choice further. Given the age of our CentOS based installer and
it's last Stable release being around a year old, we recommend 4 for new
installs.