Move downloads section name space / page type #1
When building for production we encounter a clash in directories as
although we do not publish any of this sections 'pages' individually
there is still a directory entry and associated xml file. We could
return to our pre-hugo download.html but this is far too similar to the
more intuitive downloads that we otherwise use. So move to 'dls' for
website section/page-type/content/urls.