remove animate.min.css plugin #35
To reduce our dependencies and speed up our load times. The existing
menu animation, the only place we use this now, is not worth the added
complexity. The use of this css addition is also suspected to
occasionally cause 'dead' menu links. Likely due to some emergent
property of this and our other old js and css dependencies.