Add JetBrains to 'Our Partners' section re sponsorship #84
With permission, we add the requested JetBrains logo to the
indicated home page subsection. Following on from a recent
two "All Products Pack" Licenses awarded to "The Rockstor Project".
## Includes:
- Transition 'Our Partner' logo images from static files
to assets/images, enabling our use of Hugo's image pre-processors.
Allowing us, in this case, to 'Fit' logo images as needed.

We still have an image size sensitivity here but this helps a
little with being more robust to differing sizes and formats. Note
that a future improvement could be to add a resize step to normalise
all logos before fitting them.