Update EULA page re installer #105
Drop "Rockstor Leap 15.4" nomenclature in favour of the more
representative 'Rockstor "Built on openSUSE" Leap 15.6'.

These changes are intended to normalise our reference re our base OS,
and to hopefully better adhere to the upstream guidelines as per:
- https://en.opensuse.org/openSUSE:Trademark_guidelines

N.B. in-file formatting, within EULA section, intentionally follows
that of the original, as presented to the end-user during install.

Includes
- Copyright date update.
- Specify 'Stable Updates' as Subscription type: we have others.
- Acknowledge version differences.
- Remove now redundant references.
- Minor EULA section introduction/pre-able rewording.
- Preserve underlying references to "The openSUSE Project".
- Preserve "SUSE" & "openSUSE" trademarks attribution: SUSE LLC.
- Preserve upstreams legal reference given our "Built on ..." status.